test(db): move books tests next to module
